fix build error: Type 'Uint8Array' is not assignable to type 'string'.

+++ index.ts
@@ -693,7 +693,7 @@ function isWithinCurveOrder(num: bigint): boolean {
 function normalizePrivateKey(key: PrivKey): Uint8Array {
   let num: bigint;
   if (typeof key === 'bigint' || (Number.isSafeInteger(key) && key > 0)) {
-    num = BigInt(key);
+    num = BigInt(key as bigint);
     key = num.toString(16).padStart(B32 * 2, '0');
   }
   if (typeof key === 'string') {
